requirements.txt
python-telegram-bot==20.8
google-api-python-client==2.126.0
pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.23.26
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.1
